requirements.txt
Flask
Flask-Cors
wordcloud
Pillow